+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 Date format: DD/MM/YYYY
 
-## [unreleased]
+## [unreleased]  - Fidelity
 
 - Do not override material `Theme` on `FluentApp` ([#155](https://github.com/bdlukaa/fluent_ui/pull/154))
 - Slider thumb now doesn't change inner size if hovered while disabled
@@ -11,6 +11,7 @@ Date format: DD/MM/YYYY
 - Ensure we use `Typography.body` as the default text style on `BaseButton` ([#120](https://github.com/bdlukaa/fluent_ui/issues/160))
 - **BREAKING** Renamed `standartCurve` to `standardCurve`
 - **BREAKING** Completly rework `DropDownButton`
+- Update `ButtonThemeData.uncheckedInputColor`
 
 ## [3.8.0] - Flutter Favorite - [03/02/2022]